remove map scantransponders

+++ dvb/zapit/include/zapit/sdt.h
@@ -24,7 +24,7 @@
 
 #include "types.h"
 
-unsigned long get_sdt_TsidOnid(void);
+uint32_t get_sdt_TsidOnid(void);
 int nvod_service_ids(const t_transport_stream_id, const t_original_network_id, const t_service_id, const unsigned int num, t_transport_stream_id * const, t_original_network_id * const, t_service_id * const);
 int parse_sdt(const t_transport_stream_id, const t_original_network_id, const unsigned char diseqc);
